Ship logs with a single instance of Vector, fix not shipping logs on immediate exit
Preivously, we used an instance of Vector for stdout and stderr
individually in order to determine the stream type. Now, we prefix each
stream with metadata and parse it inside of Vector in order to only run
one Vector instance.

Adding a sleep after the runc command lets Vector capture the logs. This
fixes applications not shipping logs if they exit immediately.