fix: buildQueryOptions error (#3132)
Closes FRONT-838
### TL;DR

Hide the "Instances" section in the sidebar when the data provider doesn't support query options.

### What changed?

Added a check to verify if the data provider has the `buildsQueryOptions` capability before rendering the "Instances" section in both the Subnav and CloudSidebarContent components. This prevents displaying the Instances section when it's not applicable.

### How to test?

1. Connect to a data provider that doesn't support query options
2. Verify that the "Instances" section is not displayed in the sidebar
3. Connect to a data provider that does support query options
4. Verify that the "Instances" section appears correctly

### Why make this change?

Some data providers don't support query options, making the "Instances" section irrelevant for those providers. This change improves the UI by only showing relevant navigation options based on the capabilities of the connected data provider.