Fix InlineWebSocketAdapter message delivery (#3670)
When attempting to reproduce https://www.rivet.dev/docs/actors/websocket-handler/ with createFileSystemDriver, messages sent from the actor to the client are never delivered. The connection appears to work (open/close events fire), but message routing fails silently. When using handleRawWebSocket with an InlineWebSocketAdapter, messages sent from the actor to the client are never delivered.

https://github.com/rivet-dev/rivet/blob/c9cfd90d5fb152fd47fc93acfadb1dec7f18e6bf/rivetkit-typescript/packages/rivetkit/src/actor/router-websocket-endpoints.ts#L319-L325 Looks up an adaptor, but this is never set for InlineWebSocketAdapter.