Fix InlineWebSocketAdapter message delivery (#3670)
When attempting to reproduce https://www.rivet.dev/docs/actors/websocket-handler/ with createFileSystemDriver, messages sent from the actor to the client are never delivered. The connection appears to work (open/close events fire), but message routing fails silently. When using handleRawWebSocket with an InlineWebSocketAdapter, messages sent from the actor to the client are never delivered.

https://github.com/rivet-dev/rivet/blob/e5f78d9acd1d6b5b5801b9c288e834997e701a2b/rivetkit-typescript/packages/rivetkit/src/actor/router-websocket-endpoints.ts#L319-L325 Looks up an adaptor, but this is never set for InlineWebSocketAdapter.